# Improve 2D sparse row slicing